+++ Models/Transaccion.cs
@@ -7,14 +7,14 @@ public class Transaccion
     public int _idTransaccion {get;set;}
     public int _idUsuario { get; set; }
     public int _idCategoria { get; set; }
-    public int _cantidad {get; set;}
+    public decimal _cantidad {get; set;}
     public char _tipoMovimiento {get; set;}
     public string _descripcionTransaccion {get; set;}
     public DateTime _fecTransaccion { get; set; } = DateTime.Now;
 
     public Transaccion(){}
 
-    public Transaccion (int idTransaccion, int idUsuario, int idCategoria,char tipoMovimiento, int cantidad, string descripcionTransaccion, DateTime fecTransaccion)
+    public Transaccion (int idTransaccion, int idUsuario, int idCategoria,char tipoMovimiento, decimal cantidad, string descripcionTransaccion, DateTime fecTransaccion)
     {
         _idTransaccion = idTransaccion;
         _idUsuario = idUsuario;
